Complete aplying redux tool kit
