+++ client/src/components/layouts/MyNavbar.tsx
@@ -5,20 +5,25 @@ import Button from 'react-bootstrap/Button';
 import PropTypes from 'prop-types';
 import { Container } from 'react-bootstrap';
 import { BsFillPersonFill } from 'react-icons/bs';
-import { connect, useDispatch } from 'react-redux';
-import { showModal, showRegisterModal } from '../../store/slices/auth.slice';
+import { connect, useDispatch, useSelector } from 'react-redux';
+import {
+  logOut,
+  showModal,
+  showRegisterModal
+} from '../../store/slices/auth.slice';
 import LoginForm from '../auth/LoginForm';
 import RegisterForm from '../auth/RegisterForm';
 import '../../css/landing/navbar.css';
-import store from '../../store/store';
-import { loadUser } from '../../store/actions/auth/authActions';
+import store, { RootState } from '../../store/store';
+import { loadUser } from '../../store/slices/auth.slice';
 import logoutIcon from '../../assets/logout.svg';
 
 // import { User } from "../../store/actions/types";
 
 const MyNavbar: React.FC = (props: any) => {
+  const auth = useSelector((state: RootState) => state.auth);
   const dispatch = useDispatch();
-
+  console.log(auth);
   const showLoginForm = () => {
     dispatch(showModal());
   };
@@ -27,7 +32,7 @@ const MyNavbar: React.FC = (props: any) => {
   };
 
   const logoutUser = () => {
-    //dispatch(logout());
+    dispatch(logOut());
   };
 
   useEffect(() => {
@@ -38,10 +43,10 @@ const MyNavbar: React.FC = (props: any) => {
   }, []);
 
   let navBar = null;
-  if (props.auth.user) {
+  if (auth.user) {
     navBar = (
       <>
-        <nav>Welcome {props.auth.user?.username}</nav>
+        <nav>Welcome {auth.user?.username}</nav>
         <Button
           variant="secondary"
           className="font-weight-bolder text-white"
