--- client/src/store/actions/actions.tsx
@@ -1,2 +1,0 @@
-export * from './auth/authActions';
-export * from './types';
