--- client/src/store/reducers/index.tsx
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import { authReducer } from './authReducer';
-
-const rootReducer = combineReducers({
-  auth: authReducer
-});
-
-export default rootReducer;
