+++ client/src/store/store.tsx
@@ -1,5 +1,4 @@
 import thunk from 'redux-thunk';
-import rootReducer from './reducers/index';
 import { configureStore } from '@reduxjs/toolkit';
 import authReducer from './slices/auth.slice';
